+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.4 (2020-11-09)
+------------------
+
+* added single end mode
+* extended sample sheet
+* allow mixed barcode lengths
+
 0.1.3 (2020-11-03)
 ------------------